% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survsamp.R
\name{survsamp.cluster}
\alias{survsamp.cluster}
\title{A function for surveys using cluster random sampling}
\usage{
survsamp.cluster(y = NULL, B = NULL, a = NULL, N = NULL, m = NULL,
  M = NULL, p = NULL, sig2 = NULL, sample.size.det = FALSE,
  estimate = "mean", alpha = 0.05)
}
\arguments{
\item{y}{The total of all observations in the sample needed for mean and total estimation.}

\item{B}{The desired bound on the estimation. Required for sample size determination.}

\item{a}{The total number of elements in cluster with the characteristic of interest. Needed for estimating the proportion.}

\item{N}{The number of clusters in the population. Required.}

\item{m}{The number of elements per sampled cluster.}

\item{M}{The number of elements in the population.}

\item{p}{The population proportion.}

\item{sig2}{The population variance.}

\item{sample.size.det}{Logical statement indicating if the action should be sample size determination. If false, estimation occurs.}

\item{estimate}{Indicate the desired statistic to be estimated. Options include: mean, total, and proportion.}

\item{alpha}{Significance level.}
}
\description{
This function allows you to make estimation or sample size determinations for surveys using cluster random sampling.
}
\examples{
survsamp.cluster(N=12,M=100,sig2=35000,estimate="mean",
   B=9,sample.size.det=TRUE)
set.seed(123019)
m=rpois(9,10)
survsamp.cluster(y=rpois(9,100),m=m,N=12,M=100,estimate="total")
survsamp.cluster(a=m-rpois(9,3),m=m,N=12,M=100,estimate="proportion")
}
\keyword{cluster}
\keyword{determination}
\keyword{elementary}
\keyword{estimation}
\keyword{random}
\keyword{sample}
\keyword{sampling}
\keyword{size}
\keyword{survey}
