% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survsamp.R
\name{survsamp.diff}
\alias{survsamp.diff}
\title{A function for surveys using simple random sampling when estimation of the difference between two groups is of interest}
\usage{
survsamp.diff(y1 = NULL, y2 = NULL, N1 = NULL, N2 = NULL,
  sig2.1 = NULL, sig2.2 = NULL, p1 = NULL, p2 = NULL,
  estimate = "mean", alpha = 0.05, prop.dpnt = FALSE)
}
\arguments{
\item{y1}{The data vector from population 1 to be used for estimation.}

\item{y2}{The data vector from population 2 to be used for estimation.}

\item{N1}{The population size for population 1. Required.}

\item{N2}{The population size for population 2. Required.}

\item{sig2.1}{The population 1 variance.}

\item{sig2.2}{The population 2 variance.}

\item{p1}{The population 1 proportion.}

\item{p2}{The population 2 proportion.}

\item{estimate}{Indicate the desired statistic to be estimated. Options include: mean, total, and proportion.}

\item{alpha}{Significance level.}

\item{prop.dpnt}{Logical statement indicating if the two proportions are dependent.}
}
\description{
This function allows you to make estimation or sample size determinations for the difference between two groups with surveys using simple random sampling.
}
\examples{
survsamp.diff(y1=rnorm(25,5,1),y2=rnorm(25,10,2),
     N1=46,N2=54,sig2.1=1,sig2.2=2,
     estimate="mean")
survsamp.diff(y1=rbinom(25,1,.75),y2=rbinom(25,1,.5),
     N1=46,N2=54,
     estimate="proportion",prop.dpnt=FALSE)
}
\keyword{determination}
\keyword{elementary}
\keyword{estimation}
\keyword{random}
\keyword{sample}
\keyword{sampling}
\keyword{simple}
\keyword{size}
\keyword{survey}
