% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survsamp.R
\name{survsamp.ratio}
\alias{survsamp.ratio}
\title{A function for surveys using simple random sampling when ratio estimation is of interest}
\usage{
survsamp.ratio(x = NULL, y = NULL, B = NULL, N = NULL, mux = NULL,
  taux = NULL, sig2 = NULL, sample.size.det = FALSE, estimate = "ratio",
  alpha = 0.05)
}
\arguments{
\item{x}{The independent data vector to be used for estimation.}

\item{y}{The dependent data vector to be used for estimation.}

\item{B}{The desired bound on the estimation. Required for sample size determination.}

\item{N}{The population size. Required for sample size determination.}

\item{mux}{The population mean of x.}

\item{taux}{The population total of x. This must be supplied if N is NULL.}

\item{sig2}{The population variance.}

\item{sample.size.det}{Logical statement indicating if the action should be sample size determination. If false, estimation occurs.}

\item{estimate}{Indicate the desired statistic to be estimated. Options include: ratio, mean, and total. This is only needed for sample size determination as all three are reported in estimation.}

\item{alpha}{Significance level.}
}
\description{
This function allows you to make estimation or sample size determinations for the ratio wtih surveys using simple random sampling.
}
\examples{
x=rnorm(50,5,1)
survsamp.ratio(x=x,y=x+rnorm(50),
estimate="ratio",mux=5,N=100)
}
\keyword{determination}
\keyword{elementary}
\keyword{estimation}
\keyword{random}
\keyword{sample}
\keyword{sampling}
\keyword{simple}
\keyword{size}
\keyword{survey}
