% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survsamp.R
\name{survsamp.srs}
\alias{survsamp.srs}
\title{A function for surveys using simple random sampling}
\usage{
survsamp.srs(y = NULL, B = NULL, N = NULL, p = NULL, sig2 = NULL,
  sample.size.det = FALSE, estimate = "mean", systematic = FALSE,
  alpha = 0.05)
}
\arguments{
\item{y}{The data vector to be used for estimation.}

\item{B}{The desired bound on the estimation. Required for sample size determination.}

\item{N}{The population size. Required.}

\item{p}{The population proportion.}

\item{sig2}{The population variance.}

\item{sample.size.det}{Logical statement indicating if the action should be sample size determination. If false, estimation occurs.}

\item{estimate}{Indicate the desired statistic to be estimated. Options include: mean, total, and proportion.}

\item{systematic}{Logical statement indicating if a systematic sample of a randomly ordered population is to occur. This results in a k value computed in the sample size determination.}

\item{alpha}{Significance level.}
}
\description{
This function allows you to make estimation or sample size determinations for surveys using simple random sampling.
}
\examples{
survsamp.srs(B=.1,N=100,p=.75,
   sample.size.det = TRUE,estimate="proportion")
survsamp.srs(y=rnorm(50,5,2),N=100)#,sig2=1)
survsamp.srs(B=75,N=100,sig2=1,estimate="total",
   sample.size.det = TRUE,systematic=TRUE)
}
\keyword{determination}
\keyword{elementary}
\keyword{estimation}
\keyword{random}
\keyword{sample}
\keyword{sampling}
\keyword{simple}
\keyword{size}
\keyword{survey}
\keyword{systematic}
