% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survsamp.R
\name{survsamp.strat}
\alias{survsamp.strat}
\title{A function for surveys using stratified random sampling}
\usage{
survsamp.strat(y = NULL, B = NULL, N = NULL, n = NULL, Ni = NULL,
  ni = NULL, pi = NULL, sig2 = NULL, sample.size.det = FALSE,
  budget = NULL, ss.det.cost = NULL, ss.det.a = NULL, estimate = "mean",
  alpha = 0.05)
}
\arguments{
\item{y}{The data to be used for estimation. Must be a list with length equal to the number of strata.}

\item{B}{The desired bound on the estimation. Required for sample size determination.}

\item{N}{The total population size.}

\item{n}{The total sample size.}

\item{Ni}{The population size per strata.}

\item{ni}{The sample size per strata.}

\item{pi}{The population proportion per strata.}

\item{sig2}{The population variance per strata unless equal variance across strata.}

\item{sample.size.det}{Logical statement indicating if the action should be sample size determination. If false, estimation occurs.}

\item{budget}{The budget for sample collection. Can be included in sample size determination.}

\item{ss.det.cost}{A vector pf cost to collect samples in a given strata. Can be included in sample size determination.}

\item{ss.det.a}{A vector of the allocation which indicates the proportion to be collected for each strats. Can be included in sample size determination.}

\item{estimate}{Indicate the desired statistic to be estimated. Options include: mean, total, and proportion.}

\item{alpha}{Significance level.}
}
\description{
This function allows you to make estimation or sample size determinations for surveys using stratified random sampling.
}
\examples{
survsamp.strat(B=50,N=100,
   Ni=c(46,54),sig2=c(1,2),
   estimate="total",sample.size.det = TRUE)
survsamp.strat(y=list(rnorm(25,5,1),rnorm(25,6,2)),
   N=100,n=50,ni=c(25,25),Ni=c(46,54),estimate="mean")
}
\keyword{determination}
\keyword{elementary}
\keyword{estimation}
\keyword{random}
\keyword{sample}
\keyword{sampling}
\keyword{size}
\keyword{stratified}
\keyword{survey}
